Ampute Data

The function generates a data set with missing values for data generated by GenData using the multivariate amputation approach. See mice::ampute() for more details.

Usage

AmputeData(data_complete, mech = "MAR", prop = 0.3, patterns = NULL)

Arguments

data_complete

Numeric matrix. Output of the GenData function or a three-column data set with complete data.

mech

Missing data mechanism.

prop

Proportion of missing data.

patterns

Numeric matrix consisting of zeroes and ones. Each row in the matrix represents a missing data pattern where 0 indicates a missing observation. If patterns = NULL, the default value is all possible missing data patterns for a data set with 3 columns.

Value

Returns a dataframe.

References

Schouten, R. M., Lugtig, P. and Vink, G. (2018). Generating missing values for simulation purposes: A multivariate amputation procedure. Journal of Statistical Computation and Simulation, 88(15), 1909--1930. doi:10.1080/00949655.2018.1491577

See also

Author

Ivan Jacob Agaloos Pesigan