Fit the Simple Mediation Model using Multiple Imputation

Fit the Simple Mediation Model using Multiple Imputation

Usage

FitModelMI(data_mi, mplus_bin)

Arguments

data_mi

List of numeric matrices. Output of the ImputeData function or a list of three-column data sets with imputed data.

mplus_bin

Character string. Path of Mplus binary.

Value

Returns a list with the following elements:

coef

Vector of pooled coefficients/parameter estimates \(\bar{\boldsymbol{\theta}}\).

vcov

Total covariance matrix \(\mathbf{V}_{\mathrm{total}}\).

vcov_tilde

Adjusted total covariance matrix \(\tilde{\mathbf{V}}_{\mathrm{total}}\).

vcov_between

Covariance between imputations \(\mathbf{V}_{\mathrm{between}}\).

vcov_within

Covariance within imputations \(\mathbf{V}_{\mathrm{within}}\).

ariv

Average relative increase in variance \(\mathrm{ARIV}\).

m

Number of imputations \(M\).

k

Number of parameters \(k\).

nu1

Numerator degrees of freedom \(\nu_1\) for \(D_1\).

nu2

Denominator degrees of freedom \(\nu_2\) for \(D_1\).

d1

\(D_1\) test statistic.

Details

Let $$M = \textrm{Number of imputations}, \quad \textrm{and}$$ $$m = \left\{ 1, 2, \cdots, M \right\}.$$ The vector of pooled coefficients/parameter estimates is given by $$ \bar{\boldsymbol{\theta}} = M^{-1} \sum_{m = 1}^{M} \hat{\boldsymbol{\theta}}_{m} . $$ The covariance within imputations is given by $$ \mathbf{V}_{\mathrm{within}} = M^{-1} \sum_{m = 1}^{M} \mathrm{Var} \left( \hat{\boldsymbol{\theta}}_{m} \right) $$ where \(\mathrm{Var} \left( \hat{\boldsymbol{\theta}}_{m} \right)\) is the parameter covariance matrix for the \(m^{\mathrm{th}}\) imputation. The covariance between imputations is given by $$ \mathbf{V}_{\mathrm{between}} = \left( M - 1 \right)^{-1} \sum_{m = 1}^{M} \left( \hat{\boldsymbol{\theta}}_{m} - \bar{\boldsymbol{\theta}} \right) \left( \hat{\boldsymbol{\theta}}_{m} - \bar{\boldsymbol{\theta}} \right)^{\prime} . $$ The total covariance matrix is given by $$ \mathbf{V}_{\mathrm{total}} = \mathbf{V}_{\mathrm{within}} + \mathbf{V}_{\mathrm{between}} + M^{-1} \mathbf{V}_{\mathrm{between}} . $$ The adjusted total covariance matrix is given by $$ \tilde{\mathbf{V}}_{\mathrm{total}} = \left( 1 + \mathrm{ARIV} \right) \mathbf{V}_{\mathrm{within}} $$ where $$ \mathrm{ARIV} = k^{-1} \left[ \left( 1 + M^{-1} \right) \mathrm{tr} \left( \mathbf{V}_{\mathrm{between}} \mathbf{V}_{\mathrm{within}}^{-1} \right) \right] $$ and \(k\) is the number of parameters.

References

Li, K. H., Raghunathan, T. E., & Rubin, D. B. (1991). Large-sample significance levels from multiply imputed data using moment-based statistics and an F reference distribution. Journal of the American Statistical Association, 86 (416), 1065–1073. doi:10.1080/01621459.1991.10475152

Rubin, D. B. (1987). Multiple imputation for nonresponse in surveys. John Wiley & Sons, Inc. doi:10.1002/9780470316696

See also

Author

Ivan Jacob Agaloos Pesigan