Simulation Results

Simulation Results

Usage

results

Format

A dataframe with 24,544 rows and 13 columns:

zero_hit

The proportion of replications where the confidence intervals contained zero.

theta_hit

The proportion of replications where the confidence intervals contained the population \(\alpha \beta\).

replications

Simulation replications.

taskid

Simulation Task ID.

tauprime

\(\tau^{\prime}\), that is, the path from \(X\) to \(Y\), adjusting for \(M\).

beta

\(\beta\), that is, the path from \(M\) to \(Y\).

alpha

\(\alpha\), that is, the path from \(X\) to \(M\).

n

Sample size.

sigmasqepsilonm

Error variance \(\sigma^{2}_{\varepsilon_{M}}\).

sigmasqepsilony

Error variance \(\sigma^{2}_{\varepsilon_{Y}}\).

alphabeta

\(\alpha \beta\), that is, the indirect effect of \(X\) on \(Y\) via \(M\).

mechanism

Missing data mechanism. "COMPLETE" for complete data, "MCAR" for missing completely at random, and "MAR" for missing at random.

proportion

Proportion of missing data (.0, .1, .2, .3).

method

Method used.

type1

Type I error rate.

power

Statistical power.

miss

Miss rate.

The methods are as follows:

MC.COMPLETE

for Monte Carlo method with maximum likelihood estimates for complete data.

MC.FIML

for Monte Carlo method with full information maximum likelihood estimates.

MC.MI

for Monte Carlo method with multiple imputation estimates.

MC.MI.ADJ

for Monte Carlo method with adjusted multiple imputation estimates.

NBBC.COMPLETE

for bias-corrected nonparametric bootstrap with maximum likelihood estimates for complete data.

NBBC.FIML

for full maximum likelihood nested within bias-corrected nonparametric bootstrap.

NBPC.COMPLETE

for percentile nonparametric bootstrap with full maximum likelihood estimates for complete.

NBPC.FIML

for full maximum likelihood nested within percentile nonparametric bootstrap.

SIG.COMPLETE

for joint-significant test for complete data.

SIG.FIML

for the joint-significant test with full maximum likelihood estimates.

SIG.MI

for joint-significant test with multiple imputation estimates.

SIG.MI.ADJ

for the joint-significant test with adjusted multiple imputation estimates.

Author

Ivan Jacob Agaloos Pesigan